requirements.txt
fastapi
python-dotenv
clerk-backend-api
psycopg[binary]
